e2e: Split step_impl.py into per-concern modules (#20)
Move shared Given/Then steps into steps_common.py and the add-specific
preview assertion into steps_add.py so follow-up E2E specs can proceed
in parallel without merge-conflicting on a single file.
