+++ e2e/step_impl/steps_add.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from getgauge.python import data_store, step
+
+
+@step("プレビューが表示される")
+def assert_preview_shown():
+    result = data_store.scenario["last_result"]
+    assert result.returncode == 0, (
+        f"exit code {result.returncode}\nstdout:\n{result.stdout}\nstderr:\n{result.stderr}"
+    )
+    assert "This is a preview" in result.stdout, (
+        f"preview marker not found in stdout:\n{result.stdout}"
+    )
